+++ func_adl/ast/__init__.py
@@ -1,4 +1,4 @@
-from ..function_simplifier import simplify_chained_calls  # NOQA
+from .function_simplifier import simplify_chained_calls  # NOQA
 from .aggregate_shortcuts import aggregate_node_transformer  # NOQA
 from .call_stack import argument_stack, stack_frame  # NOQA
 from .func_adl_ast_utils import change_extension_functions_to_calls  # NOQA
